# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mystics Inventory API — multi-tenant inventory SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customer-payments:
    get:
      operationId: listCustomerPayments
      tags: [customer-payments]
      parameters:
        - { in: query, name: customerId, required: false, schema: { type: integer } }
        - { in: query, name: mode, required: false, schema: { type: string } }
        - { in: query, name: from, required: false, schema: { type: string } }
        - { in: query, name: to, required: false, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CustomerPayment" }
    post:
      operationId: createCustomerPayment
      tags: [customer-payments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCustomerPaymentPayload" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerPaymentDetail" }
  /customer-payments/{id}:
    get:
      operationId: getCustomerPayment
      tags: [customer-payments]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerPaymentDetail" }
    delete:
      operationId: deleteCustomerPayment
      tags: [customer-payments]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "204": { description: No Content }
  /customer-payments/{id}/receipt.pdf:
    get:
      operationId: downloadCustomerPaymentReceipt
      tags: [customer-payments]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Receipt PDF binary
          content:
            application/pdf:
              schema:
                type: string
                format: binary
...
components:
  schemas:
...
    CustomerPayment:
      type: object
      properties:
        id: { type: integer }
        customerId: { type: integer }
        customerName: { type: string }
        paymentDate: { type: string }
        amount: { type: number }
        mode: { type: string }
        referenceNumber: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        bankAccountLabel: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, customerId, customerName, paymentDate, amount, mode, referenceNumber, notes, bankAccountLabel, createdAt]
    CustomerPaymentAllocation:
      type: object
      properties:
        id: { type: integer }
        paymentId: { type: integer }
        salesOrderId: { type: integer }
        salesOrderNumber: { type: string }
        salesOrderTotal: { type: number }
        salesOrderBalanceDue: { type: number }
        amount: { type: number }
      required: [id, paymentId, salesOrderId, salesOrderNumber, salesOrderTotal, salesOrderBalanceDue, amount]
    CustomerPaymentDetail:
      type: object
      properties:
        payment: { $ref: "#/components/schemas/CustomerPayment" }
        allocations:
          type: array
          items: { $ref: "#/components/schemas/CustomerPaymentAllocation" }
      required: [payment, allocations]
    CustomerPaymentAllocationInput:
      type: object
      properties:
        salesOrderId: { type: integer }
        amount: { type: number }
      required: [salesOrderId, amount]
    CreateCustomerPaymentPayload:
      type: object
      properties:
        customerId: { type: integer }
        paymentDate: { type: string }
        amount: { type: number }
        mode: { type: string }
        referenceNumber: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        bankAccountLabel: { type: ["string", "null"] }
        allocations:
          type: array
          items: { $ref: "#/components/schemas/CustomerPaymentAllocationInput" }
      required: [customerId, amount, mode]
...